requirements_dev.txt
pytest
pytest-django
pep257
selenium
model_mommy
isort
requests
flake8
pep8-naming